// ui/src/api/generated/models/plugin-template-checklist-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.24.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 模板初始化检查项
 * @export
 * @interface PluginTemplateChecklistItem
 */
export interface PluginTemplateChecklistItem {
  /**
   * 受众
   * @type {string}
   * @memberof PluginTemplateChecklistItem
   */
  audience?: string
  /**
   * 说明
   * @type {string}
   * @memberof PluginTemplateChecklistItem
   */
  description?: string
  /**
   * 检查项键
   * @type {string}
   * @memberof PluginTemplateChecklistItem
   */
  key?: string
  /**
   * 状态
   * @type {string}
   * @memberof PluginTemplateChecklistItem
   */
  status?: string
  /**
   * 标题
   * @type {string}
   * @memberof PluginTemplateChecklistItem
   */
  title?: string
}