// ui/src/api/generated/models/plugin-template-feature-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.24.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 模板功能项
 * @export
 * @interface PluginTemplateFeatureItem
 */
export interface PluginTemplateFeatureItem {
  /**
   * 区域
   * @type {string}
   * @memberof PluginTemplateFeatureItem
   */
  area?: string
  /**
   * 说明
   * @type {string}
   * @memberof PluginTemplateFeatureItem
   */
  description?: string
  /**
   * 是否默认启用
   * @type {boolean}
   * @memberof PluginTemplateFeatureItem
   */
  enabled?: boolean
  /**
   * 功能键
   * @type {string}
   * @memberof PluginTemplateFeatureItem
   */
  key?: string
  /**
   * 标题
   * @type {string}
   * @memberof PluginTemplateFeatureItem
   */
  title?: string
}