// ui/src/api/generated/models/plugin-template-overview.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.24.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { PluginTemplateChecklistItem } from './plugin-template-checklist-item'
// May contain unused imports in some cases
// @ts-ignore
import type { PluginTemplateFeatureItem } from './plugin-template-feature-item'
// May contain unused imports in some cases
// @ts-ignore
import type { PluginTemplateStatItem } from './plugin-template-stat-item'

/**
 * 模板概览信息
 * @export
 * @interface PluginTemplateOverview
 */
export interface PluginTemplateOverview {
  /**
   * 主色
   * @type {string}
   * @memberof PluginTemplateOverview
   */
  accentColor?: string
  /**
   * 当前视角
   * @type {string}
   * @memberof PluginTemplateOverview
   */
  audience?: string
  /**
   * 视角标签
   * @type {string}
   * @memberof PluginTemplateOverview
   */
  audienceLabel?: string
  /**
   * 初始化清单
   * @type {Array<PluginTemplateChecklistItem>}
   * @memberof PluginTemplateOverview
   */
  checklist?: Array<PluginTemplateChecklistItem>
  /**
   * ConfigMap 名称
   * @type {string}
   * @memberof PluginTemplateOverview
   */
  configMapName?: string
  /**
   * Console 路由
   * @type {string}
   * @memberof PluginTemplateOverview
   */
  consolePath?: string
  /**
   * 页面密度
   * @type {string}
   * @memberof PluginTemplateOverview
   */
  density?: string
  /**
   * 显示名
   * @type {string}
   * @memberof PluginTemplateOverview
   */
  displayName?: string
  /**
   * 是否启用附件扩展标签页
   * @type {boolean}
   * @memberof PluginTemplateOverview
   */
  enableAttachmentProvider?: boolean
  /**
   * 是否启用 Console 页面
   * @type {boolean}
   * @memberof PluginTemplateOverview
   */
  enableConsoleDashboard?: boolean
  /**
   * 是否启用 UC 页面
   * @type {boolean}
   * @memberof PluginTemplateOverview
   */
  enableUcDashboard?: boolean
  /**
   * 功能矩阵
   * @type {Array<PluginTemplateFeatureItem>}
   * @memberof PluginTemplateOverview
   */
  features?: Array<PluginTemplateFeatureItem>
  /**
   * 生成时间
   * @type {string}
   * @memberof PluginTemplateOverview
   */
  generatedAt?: string
  /**
   * 生成客户端输出目录
   * @type {string}
   * @memberof PluginTemplateOverview
   */
  generatedClientPath?: string
  /**
   * 插件名
   * @type {string}
   * @memberof PluginTemplateOverview
   */
  pluginName?: string
  /**
   * Setting 名称
   * @type {string}
   * @memberof PluginTemplateOverview
   */
  settingName?: string
  /**
   * 统计信息
   * @type {Array<PluginTemplateStatItem>}
   * @memberof PluginTemplateOverview
   */
  stats?: Array<PluginTemplateStatItem>
  /**
   * 支持文档链接
   * @type {string}
   * @memberof PluginTemplateOverview
   */
  supportLink?: string
  /**
   * UC 路由
   * @type {string}
   * @memberof PluginTemplateOverview
   */
  ucPath?: string
}