// ui/src/api/generated/models/plugin-template-record.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.24.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { Metadata } from './metadata'

/**
 *
 * @export
 * @interface PluginTemplateRecord
 */
export interface PluginTemplateRecord {
  /**
   *
   * @type {string}
   * @memberof PluginTemplateRecord
   */
  apiVersion: string
  /**
   * 创建时间
   * @type {string}
   * @memberof PluginTemplateRecord
   */
  createTime?: string
  /**
   * 描述
   * @type {string}
   * @memberof PluginTemplateRecord
   */
  description?: string
  /**
   * 业务主键
   * @type {string}
   * @memberof PluginTemplateRecord
   */
  id: string
  /**
   *
   * @type {string}
   * @memberof PluginTemplateRecord
   */
  kind: string
  /**
   *
   * @type {Metadata}
   * @memberof PluginTemplateRecord
   */
  metadata: Metadata
  /**
   * 状态
   * @type {string}
   * @memberof PluginTemplateRecord
   */
  status: string
  /**
   * 标题
   * @type {string}
   * @memberof PluginTemplateRecord
   */
  title: string
  /**
   * 更新时间
   * @type {string}
   * @memberof PluginTemplateRecord
   */
  updateTime?: string
}