// ui/src/api/generated/models/plugin-template-record-create-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.24.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 示例数据记录创建请求
 * @export
 * @interface PluginTemplateRecordCreateRequest
 */
export interface PluginTemplateRecordCreateRequest {
  /**
   * 描述
   * @type {string}
   * @memberof PluginTemplateRecordCreateRequest
   */
  description?: string
  /**
   * 业务主键，留空时自动生成
   * @type {string}
   * @memberof PluginTemplateRecordCreateRequest
   */
  id?: string
  /**
   * 状态
   * @type {string}
   * @memberof PluginTemplateRecordCreateRequest
   */
  status?: string
  /**
   * 标题
   * @type {string}
   * @memberof PluginTemplateRecordCreateRequest
   */
  title: string
}