// ui/src/api/generated/models/plugin-template-record-update-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.24.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 示例数据记录更新请求
 * @export
 * @interface PluginTemplateRecordUpdateRequest
 */
export interface PluginTemplateRecordUpdateRequest {
  /**
   * 描述
   * @type {string}
   * @memberof PluginTemplateRecordUpdateRequest
   */
  description?: string
  /**
   * 状态
   * @type {string}
   * @memberof PluginTemplateRecordUpdateRequest
   */
  status?: string
  /**
   * 标题
   * @type {string}
   * @memberof PluginTemplateRecordUpdateRequest
   */
  title: string
}