// ui/src/api/generated/models/plugin-template-stat-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.24.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 模板统计卡片
 * @export
 * @interface PluginTemplateStatItem
 */
export interface PluginTemplateStatItem {
  /**
   * 提示
   * @type {string}
   * @memberof PluginTemplateStatItem
   */
  helper?: string
  /**
   * 统计项键
   * @type {string}
   * @memberof PluginTemplateStatItem
   */
  key?: string
  /**
   * 标题
   * @type {string}
   * @memberof PluginTemplateStatItem
   */
  label?: string
  /**
   * 视觉语义
   * @type {string}
   * @memberof PluginTemplateStatItem
   */
  tone?: string
  /**
   * 值
   * @type {string}
   * @memberof PluginTemplateStatItem
   */
  value?: string
}